Add other movie data fragments
