+++ app/src/main/java/com/sheilambadi/android/retrofitproject/fragments/LatestFragment.java
@@ -1,0 +1,30 @@
+package com.sheilambadi.android.retrofitproject.fragments;
+
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.sheilambadi.android.retrofitproject.R;
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class LatestFragment extends Fragment {
+
+
+    public LatestFragment() {
+        // Required empty public constructor
+    }
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_latest, container, false);
+    }
+
+}
