+++ app/src/main/res/layout/fragment_group_movies.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.GroupMoviesFragment">
+    tools:context=".fragments.BrowseMoviesFragment">
 
     <android.support.design.widget.CoordinatorLayout
         android:layout_width="match_parent"
